backend/bak_requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.11.0
python-dotenv==1.1.0
psycopg[binary]==3.3.3
gunicorn==23.0.0
pytest==8.3.5
Pillow==10.4.0
pytesseract==0.3.13
requests==2.32.5
pandas==2.2.3
numpy==1.26.4
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pillow-heif